+++ packages/svelte/CHANGELOG.md
@@ -6,6 +6,11 @@ description: All notable changes will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- **Format**: Added `Format.Time` for formatting time values and examples for basic, date input, seconds, custom AM/PM
+  labels, and locale
+
 ### Fixed
 
 - **Build**: Fixed CJS build issues in downstream Zag.js packages